tests/dynamorio: add a new DR based tester for EMFILE handling
DR is suitable for testing EMFILE handling because it can open files
transparently and consume fds. This patch adds a new DR client for
this purpose. We can enhance this client for inserting fd consuming at
arbital timing.

This patch also adds a test scinario for testing sockfd_shrink()
(01.sh). With this script, we can confirm that sockfd_shrink() is
actually called and does it work.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>