lib: move sockfd cache from sheep to lib
On some subcommands, collie also issues lots of request to sheeps. So
collie can enjoy sockfd caching. For this purpose, this patch moves
sockfd from sheep to lib and generalize the interfaces of sockfd.

This patch doesn't change anything related to IO NIC in sockfd
cache. Because collie can know address and port of IO NIC so
potentially use them.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>